Export the implementation of the ws_connection_processor
